chore(fastqc): minor review polish (PR #420)
Non-blocking follow-ups from the latest review round:

- lib.rs: widen the random temp-table suffix from u32 to u64 for the
  fastqc and pileup UDTF registrations (larger collision-free space; the
  two paths the review called out).
- fastqc_op.py: document that ALL_MODULES must stay in sync with the
  upstream Rust ModuleSet enum (source of truth), noting the KeyError
  failure mode if it drifts.
- benchmark-scripts/correctness_verify.sh: drop stale "#212 / NEW local"
  development notes from the header comment.

+++ src/lib.rs
@@ -811,7 +811,7 @@ fn py_register_pileup_table(
                     };
 
                 let depth_provider = pileup::DepthTableProvider::new(provider, config);
-                let table_name = format!("_pileup_{}", rand::random::<u32>());
+                let table_name = format!("_pileup_{}", rand::random::<u64>());
                 ctx.register_table(&table_name, Arc::new(depth_provider))
                     .map_err(|e| format!("Failed to register depth table: {}", e))?;
                 // No execution — just register and return the name
@@ -854,7 +854,7 @@ fn py_register_fastqc_table(
         .map_err(|e| PyRuntimeError::new_err(format!("Failed to create FASTQ provider: {e}")))?;
 
         let provider = crate::fastqc::FastqcTableProvider::new(Arc::new(fq), modules);
-        let table_name = format!("_fastqc_{}", rand::random::<u32>());
+        let table_name = format!("_fastqc_{}", rand::random::<u64>());
         ctx.register_table(&table_name, Arc::new(provider))
             .map_err(|e| {
                 PyRuntimeError::new_err(format!("Failed to register fastqc table: {e}"))
